chore: add dry run option to update_native_dependencies

+++ update_native_dependencies.py
@@ -1,3 +1,4 @@
+import argparse
 import hashlib
 import re
 import subprocess
@@ -23,7 +24,7 @@ def _sha256(url):
     return m.hexdigest()
 
 
-def _update_cpython():
+def _update_cpython(dry_run):
     dockerfile = Path(__file__).parent / "docker" / "Dockerfile"
     lines = dockerfile.read_text().splitlines()
     re_ = re.compile(r"^RUN.*/build-cpython.sh (?P<version>.*)$")
@@ -38,13 +39,14 @@ def _update_cpython():
             url = f"https://www.python.org/ftp/python/{latest_version.major}.{latest_version.minor}.{latest_version.micro}"
             _sha256(f"{url}/{root}.tgz")
             lines[i] = lines[i].replace(match["version"], str(latest_version))
-            dockerfile.write_text("\n".join(lines) + "\n")
             message = f"Bump CPython {current_version} → {latest_version}"
             print(message)
-            subprocess.check_call(["git", "commit", "-am", message])
+            if not dry_run:
+                dockerfile.write_text("\n".join(lines) + "\n")
+                subprocess.check_call(["git", "commit", "-am", message])
 
 
-def _update_with_root(tool):
+def _update_with_root(tool, dry_run):
     repo = {
         "autoconf": "autotools-mirror/autoconf",
         "automake": "autotools-mirror/automake",
@@ -72,14 +74,15 @@ def _update_with_root(tool):
             sha256 = _sha256(f"{url}/{root}.tar.gz")
             lines[i + 0] = f"RUN export {tool.upper()}_ROOT={root} && \\"
             lines[i + 1] = f"    export {tool.upper()}_HASH={sha256} && \\"
-            dockerfile.write_text("\n".join(lines) + "\n")
             message = f"Bump {tool} {current_version} → {latest_version}"
             print(message)
-            subprocess.check_call(["git", "commit", "-am", message])
+            if not dry_run:
+                dockerfile.write_text("\n".join(lines) + "\n")
+                subprocess.check_call(["git", "commit", "-am", message])
         break
 
 
-def _update_sqlite():
+def _update_sqlite(dry_run):
     dockerfile = Path(__file__).parent / "docker" / "Dockerfile"
     lines = dockerfile.read_text().splitlines()
     re_ = re.compile(f"^RUN export SQLITE_AUTOCONF_ROOT=sqlite-autoconf-(?P<version>\\S+) && \\\\$")
@@ -104,14 +107,15 @@ def _update_sqlite():
             lines[i + 0] = f"RUN export SQLITE_AUTOCONF_ROOT={root} && \\"
             lines[i + 1] = f"    export SQLITE_AUTOCONF_HASH={sha256} && \\"
             lines[i + 2] = f"    export SQLITE_AUTOCONF_DOWNLOAD_URL={url} && \\"
-            dockerfile.write_text("\n".join(lines) + "\n")
             message = f"Bump sqlite {current_version} → {latest_version}"
             print(message)
-            subprocess.check_call(["git", "commit", "-am", message])
+            if not dry_run:
+                dockerfile.write_text("\n".join(lines) + "\n")
+                subprocess.check_call(["git", "commit", "-am", message])
         break
 
 
-def _update_with_gh(tool):
+def _update_with_gh(tool, dry_run):
     repo = {
         "patchelf": "NixOS/patchelf",
         "libxcrypt": "besser82/libxcrypt",
@@ -131,20 +135,24 @@ def _update_with_gh(tool):
             sha256 = _sha256(f"{url}/{latest_tag}.tar.gz")
             lines[i + 0] = f"RUN export {tool.upper()}_VERSION={latest_version} && \\"
             lines[i + 1] = f"    export {tool.upper()}_HASH={sha256} && \\"
-            dockerfile.write_text("\n".join(lines) + "\n")
             message = f"Bump {tool} {current_version} → {latest_version}"
             print(message)
-            subprocess.check_call(["git", "commit", "-am", message])
+            if not dry_run:
+                dockerfile.write_text("\n".join(lines) + "\n")
+                subprocess.check_call(["git", "commit", "-am", message])
         break
 
 
 def main():
-    _update_cpython()
-    _update_sqlite()
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--dry-run", dest="dry_run", action="store_true", help="dry run")
+    args = parser.parse_args()
+    _update_cpython(args.dry_run)
+    _update_sqlite(args.dry_run)
     for tool in ["autoconf", "automake", "libtool", "git", "swig", "openssl"]:
-        _update_with_root(tool)
+        _update_with_root(tool, args.dry_run)
     for tool in ["patchelf", "libxcrypt"]:
-        _update_with_gh(tool)
+        _update_with_gh(tool, args.dry_run)
 
 
 if __name__ == "__main__":
